clean up and reorg
